Move clear pending to manuscript helper and add unit test for function and ensure its called incontroller
