+++ server/xpub-controller/manuscript.js
@@ -1,6 +1,5 @@
 const logger = require('@pubsweet/logger')
 const ManuscriptModel = require('@elifesciences/xpub-model').Manuscript
-const FileModel = require('@elifesciences/xpub-model').File
 const Notification = require('./notification')
 const { FilesHelper, ManuscriptHelper } = require('./helpers')
 
@@ -18,24 +17,6 @@ class Manuscript {
       this.storage,
       this.filesHelper,
     )
-  }
-
-  static async clearPendingFile(manuscript) {
-    const pendingFileIndex = manuscript.files.findIndex(
-      element => element.type === 'MANUSCRIPT_SOURCE_PENDING',
-    )
-
-    if (pendingFileIndex >= 0) {
-      const pendingFile = await FileModel.find(
-        manuscript.files[pendingFileIndex].id,
-      )
-      await pendingFile.delete()
-      logger.info(
-        `Pending file removed ${pendingFileIndex} ${
-          manuscript.files[pendingFileIndex].filename
-        } | ${manuscript.id}`,
-      )
-    }
   }
 
   async upload(manuscriptId, file, fileSize) {
@@ -74,6 +55,8 @@ class Manuscript {
         progress,
         manuscriptId,
       )
+      const manuscript = await ManuscriptModel.find(manuscriptId, this.userId)
+      await ManuscriptHelper.clearPendingFile(manuscript)
       throw error
     }
